a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,10 +7,10 @@
 */
 int main(void)
 {
-int i, prev = 1, curr = 2, next;
+long long int i, prev = 1, curr = 2, next;
 for (i = 1 ; i <= 50 ; i++)
 {
-printf("%d", prev);
+printf("%lld", prev);
 if (i != 50)
 {
 printf(", ");
